Removed TabStack, simplified routing
