--- frontend/navigation/ProfileStack.tsx
@@ -1,14 +1,0 @@
-import React from 'react';
-import { createStackNavigator } from '@react-navigation/stack';
-import ProfileScreen from '../screens/ProfileScreen';
-
-
-const Stack = createStackNavigator();
-
-export default function ProfileStack() {
-  return (
-    <Stack.Navigator>
-      <Stack.Screen name='Profile' component={ProfileScreen} />
-    </Stack.Navigator>
-  );
-}
